docs/notebook/requirements.txt
# [dev-packages]
pytest==6.2.5
pytest-cov==4.1.0
autopep8==2.0.4

# [packages]
jupyterlab==3.0.16
ipywidgets>=7.6.3,<=8.0.2
matplotlib-venn==0.11.10
piexif==1.1.3
pickleshare==0.7.5
pandas>=1.1.5,<2.0.0
